backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.31.0
anthropic==0.28.0
reportlab==4.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
